+++ Plugin.cs
@@ -28,7 +28,7 @@ public class VileVendingMachinePlugin : BaseUnityPlugin
 {
     public const string ModGuid = $"LCM_VileVendingMachine|{ModVersion}";
     private const string ModName = "Lethal Company Vile Vending Machine Mod";
-    private const string ModVersion = "1.1.1";
+    private const string ModVersion = "1.1.2";
 
     private readonly Harmony _harmony = new(ModGuid);
 
@@ -68,6 +68,7 @@ public class VileVendingMachinePlugin : BaseUnityPlugin
         
         _harmony.PatchAll(typeof(VendingMachineRegistry));
         _harmony.PatchAll(typeof(VileVendingMachinePlugin));
+        _harmony.PatchAll(typeof(CompanyColaBehaviour));
         Mls.LogInfo($"Plugin {ModName} is loaded!");
     }
 
@@ -155,35 +156,53 @@ public class VileVendingMachinePlugin : BaseUnityPlugin
         }
     }
     
-    private static void RegisterEnemyWithConfig(bool enemyEnabled, string configMoonRarity, EnemyType enemy, TerminalNode terminalNode, TerminalKeyword terminalKeyword) {
-        if (enemyEnabled) { 
+    private static void RegisterEnemyWithConfig(bool enemyEnabled, string configMoonRarity, EnemyType enemy, TerminalNode terminalNode, TerminalKeyword terminalKeyword) 
+    {
+        if (enemyEnabled) 
+        { 
             (Dictionary<Levels.LevelTypes, int> spawnRateByLevelType, Dictionary<string, int> spawnRateByCustomLevelType) = ConfigParsing(configMoonRarity);
             Enemies.RegisterEnemy(enemy, spawnRateByLevelType, spawnRateByCustomLevelType, terminalNode, terminalKeyword);
-                
-        } else {
+        } 
+        else 
+        {
             Enemies.RegisterEnemy(enemy, 0, Levels.LevelTypes.All, terminalNode, terminalKeyword);
         }
     }
-
-    // Got from the giant specimens mod
-    private static (Dictionary<Levels.LevelTypes, int> spawnRateByLevelType, Dictionary<string, int> spawnRateByCustomLevelType) ConfigParsing(string configMoonRarity) {
+    
+    private static (Dictionary<Levels.LevelTypes, int> spawnRateByLevelType, Dictionary<string, int> spawnRateByCustomLevelType) ConfigParsing(string configMoonRarity) 
+    {
         Dictionary<Levels.LevelTypes, int> spawnRateByLevelType = new();
         Dictionary<string, int> spawnRateByCustomLevelType = new();
-        foreach (string entry in configMoonRarity.Split(',').Select(s => s.Trim())) {
+        foreach (string entry in configMoonRarity.Split(',').Select(s => s.Trim())) 
+        {
             string[] entryParts = entry.Split(':');
 
             if (entryParts.Length != 2) continue;
             string name = entryParts[0];
             if (!int.TryParse(entryParts[1], out int spawnrate)) continue;
 
-            if (Enum.TryParse(name, true, out Levels.LevelTypes levelType)) {
+            if (Enum.TryParse(name, true, out Levels.LevelTypes levelType)) 
+            {
                 spawnRateByLevelType[levelType] = spawnrate;
                 Mls.LogDebug($"Registered spawn rate for level type {levelType} to {spawnrate}");
-            } else {
-                spawnRateByCustomLevelType[name] = spawnrate;
-                Mls.LogDebug($"Registered spawn rate for custom level type {name} to {spawnrate}");
+            } 
+            else 
+            {
+                // Try appending "Level" to the name and re-attempt parsing
+                string modifiedName = name + "Level";
+                if (Enum.TryParse(modifiedName, true, out levelType))
+                {
+                    spawnRateByLevelType[levelType] = spawnrate;
+                    Mls.LogDebug($"Registered spawn rate for level type {levelType} to {spawnrate}");
+                }
+                else
+                {
+                    spawnRateByCustomLevelType[name] = spawnrate;
+                    Mls.LogDebug($"Registered spawn rate for custom level type {name} to {spawnrate}");
+                }
             }
         }
+        
         return (spawnRateByLevelType, spawnRateByCustomLevelType);
     }
 }
@@ -345,7 +364,8 @@ public static class LobbyCompatibilityChecker
     public static bool Enabled => BepInEx.Bootstrap.Chainloader.PluginInfos.ContainsKey("BMX.LobbyCompatibility");
 
     [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
-    public static void Init() {
+    public static void Init() 
+    {
         PluginHelper.RegisterPlugin(PluginInfo.PLUGIN_GUID, Version.Parse(PluginInfo.PLUGIN_VERSION), CompatibilityLevel.Everyone, VersionStrictness.Patch);
     }
 }